Refactor: restructured files
