+++ packages/tangle/storage_test.go
@@ -8,6 +8,27 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func TestAttachment(t *testing.T) {
+	transactionID, err := ledgerstate.TransactionIDFromRandomness()
+	assert.NoError(t, err)
+	// TODO: maybe make this public?
+	messageID := randomMessageID()
+
+	attachment := NewAttachment(transactionID, messageID)
+
+	assert.Equal(t, transactionID, attachment.TransactionID())
+	assert.Equal(t, messageID, attachment.MessageID())
+
+	clonedAttachment, consumedBytes, err := AttachmentFromBytes(attachment.Bytes())
+	if err != nil {
+		panic(err)
+	}
+
+	assert.Equal(t, AttachmentLength, consumedBytes)
+	assert.Equal(t, transactionID, clonedAttachment.TransactionID())
+	assert.Equal(t, messageID, clonedAttachment.MessageID())
+}
+
 func TestStorage_StoreAttachment(t *testing.T) {
 	tangle := New()
 	defer tangle.Shutdown()
